refactor mma test
